I tried many methods
and reached a max test accuracy of 75%. Not happy with it at all but I just realised that Kaggle won't give me a medal even if I get 99% test accuracy in this problem since it is an expired competition. I am moving on to greener pastures , that is, an active competition.